requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
langchain 
faiss-cpu 
sentence-transformers 
langchain-community
